+++ builds/compromise.js
@@ -9,7 +9,7 @@ module.exports={
   "author": "Spencer Kelly <spencermountain@gmail.com> (http://spencermounta.in)",
   "name": "compromise",
   "description": "natural language processing in the browser",
-  "version": "11.10.1",
+  "version": "11.11.0",
   "main": "./builds/compromise.js",
   "types": "./compromise.d.ts",
   "repository": {
@@ -18,6 +18,7 @@ module.exports={
   },
   "scripts": {
     "test": "tape \"./test/unit/**/*.test.js\" | tap-dancer",
+    "test:spec": "tape \"./test/unit/**/*.test.js\" | tap-spec",
     "testb": "TESTENV=prod tape \"./test/unit/**/*.test.js\" | tap-dancer",
     "buildTest": "TESTENV=prod node ./scripts/test.js",
     "test:types": "tsc --project test/types",
@@ -6589,29 +6590,32 @@ var quotemarks = {
 };
 
 // Open quote match black list.
-var blacklist = ['twas'];
+// const blacklist = [
+//   'twas'
+// ];
 
 // Convert the close quote to a regex.
 Object.keys(quotemarks).forEach(function (open) {
-  quotemarks[open].regex = new RegExp(quotemarks[open].close + '[;:,.]*$');
+  quotemarks[open].regex = new RegExp(quotemarks[open].close + '[;:,.]*');
   quotemarks[open].open = open;
 });
 
 // Improve open match detection.
-var startQuote = new RegExp('^[' + Object.keys(quotemarks).join('') + ']+' + '(?!' + blacklist.join('|') + ')');
+var startQuote = new RegExp('[' + Object.keys(quotemarks).join('') + ']'
+// '(?!' + blacklist.join('|') + ')'
+);
 
 //tag a inline quotation as such
 var quotation_step = function quotation_step(ts) {
   // Isolate the text so it doesn't change.
-  var terms = ts.terms.slice(0).map(function (e) {
-    return e.text;
-  });
+  var terms = ts.terms.slice(0); //.map(e => e.text);
 
   var _loop = function _loop(i) {
+
     var t = ts.terms[i];
-    if (startQuote.test(t.text)) {
+    if (startQuote.test(t.whitespace.before)) {
       // Get the match and split it into groups
-      var quotes = t.text.match(startQuote).shift().split('');
+      var quotes = t.whitespace.before.match(startQuote).shift().split('');
       // Get close and tag info.
       quotes = quotes.map(function (mark) {
         return quotemarks[mark];
@@ -6620,25 +6624,26 @@ var quotation_step = function quotation_step(ts) {
 
       var _loop2 = function _loop2(_o) {
         // max-length don't go-on forever
-        if (!ts.terms[i + _o] || _o > 28) {
+        if (!terms[i + _o] || _o > 28) {
           return 'break';
         }
         // Find the close.
         var index = quotes.findIndex(function (q) {
-          return q.regex.test(terms[i + _o]);
+          return q.regex.test(terms[i + _o].whitespace.after);
         });
         if (index !== -1) {
           // Remove the found
           var quote = quotes.splice(index, 1).pop();
-          terms[i + _o] = terms[i + _o].replace(quote.regex, '');
+          // terms[i + o].whitespace.after = terms[i + o].whitespace.after.replace(quote.regex, '');
 
           if (quote.regex.test(ts.terms[i + _o].normal)) {
-            ts.terms[i + _o].normal.replace(quote.regex, '');
+            ts.terms[i + _o].whitespace.after.replace(quote.regex, '');
           }
           // Tag the things.
           t.tag('StartQuotation', 'quotation_open');
           ts.terms[i + _o].tag('EndQuotation', 'quotation_close');
-          ts.slice(i, i + _o + 1).tag(quote.tag, 'quotation_step');
+          ts.slice(i, i + _o + 1).tag('Quotation', 'quotation_step');
+          // ts.slice(i, i + o + 1).tag(quote.tag, 'quotation_step');
           // Compensate for multiple close quotes ('"Really"')
           _o -= 1;
           if (!quotes.length) {
@@ -6659,6 +6664,14 @@ var quotation_step = function quotation_step(ts) {
   for (var i = 0; i < terms.length; i++) {
     _loop(i);
   } // for all terms
+
+  //fix any issues post-process
+  if (ts.has('#StartQuotation') === true && ts.has('#EndQuotation') === false) {
+    // ts.unTag('Quotation');
+  }
+  if (ts.has('#EndQuotation') === true && ts.has('#StartQuotation') === false) {
+    // ts.unTag('Quotation');
+  }
   return ts;
 };
 module.exports = quotation_step;
@@ -6687,7 +6700,7 @@ var apostrophes = '\'‘’‛‚‵′`´';
 
 // [^\w]* match 0 or more of any char that is NOT alphanumeric
 var afterWord = new RegExp('([a-z]s[' + apostrophes + '])\\W*$');
-var apostrophe = new RegExp('s?[' + apostrophes + ']s?$');
+var hasApostrophe = new RegExp('[' + apostrophes + ']');
 var trailers = new RegExp('[^' + apostrophes + '\\w]+$');
 
 //these are always contractions
@@ -6713,13 +6726,14 @@ var is_possessive = function is_possessive(terms, text, index) {
     return m;
   });
   // If no apostrophe s or s apostrophe
-  var hasApostrophe = apostrophe.test(text);
+  var endTick = hasApostrophe.test(thisWord.whitespace.after);
   // "spencers'" - this is always possessive - eg "flanders'"
   var hasPronoun = thisWord.tags.Pronoun;
 
-  if (inBlacklist || hasPronoun || !hasApostrophe) {
+  if (inBlacklist || hasPronoun || !endTick) {
     return false;
   }
+  console.log(text);
   if (afterWord.test(text) || nextWord === undefined) {
     return true;
   }
@@ -6765,7 +6779,6 @@ var possessiveStep = function possessiveStep(ts) {
 
     // Post checking for quotes. e.g: Carlos'. -> Carlos'
     text = text.replace(trailers, '');
-
     if (is_possessive(ts, text, i)) {
       // If it's not already a noun, co-erce it to one
       if (!term.tags['Noun']) {
@@ -8379,57 +8392,14 @@ module.exports = {
   EndQuotation: {
     isA: 'Quotation'
   },
-  StraightDoubleQuotes: {
-    isA: 'Quotation'
-  },
-  StraightDoubleQuotesWide: {
-    isA: 'Quotation'
-  },
-  StraightSingleQuotes: {
-    isA: 'Quotation'
-  },
-  CommaDoubleQuotes: {
-    isA: 'Quotation'
-  },
-  CommaSingleQuotes: {
-    isA: 'Quotation'
-  },
-  CurlyDoubleQuotesReversed: {
-    isA: 'Quotation'
-  },
-  CurlySingleQuotesReversed: {
-    isA: 'Quotation'
-  },
-  LowCurlyDoubleQuotes: {
-    isA: 'Quotation'
-  },
-  LowCurlyDoubleQuotesReversed: {
-    isA: 'Quotation'
-  },
-  LowCurlySingleQuotes: {
-    isA: 'Quotation'
-  },
-  AngleDoubleQuotes: {
-    isA: 'Quotation'
-  },
-  AngleSingleQuotes: {
-    isA: 'Quotation'
-  },
-  PrimeSingleQuotes: {
-    isA: 'Quotation'
-  },
-  PrimeDoubleQuotes: {
-    isA: 'Quotation'
-  },
-  PrimeTripleQuotes: {
-    isA: 'Quotation'
-  },
-  LowPrimeDoubleQuotesReversed: {
-    isA: 'Quotation'
-  },
   //parentheses
-  EndBracket: {},
-  StartBracket: {}
+  Parentheses: {},
+  EndBracket: {
+    isA: 'Parentheses'
+  },
+  StartBracket: {
+    isA: 'Parentheses'
+  }
 };
 
 },{}],140:[function(_dereq_,module,exports){
@@ -9434,11 +9404,32 @@ module.exports = {
 
 },{"../fns":3,"../log":6}],162:[function(_dereq_,module,exports){
 'use strict';
-//punctuation regs-
+// const quotes = [ //
+//   ['"', '"'],
+//   ['\u0022', '\u0022'],
+//   ['\uFF02', '\uFF02'],
+//   ['\u0027', '\u0027'],
+//   ['\u201C', '\u201D'],
+//   ['\u2018', '\u2019'],
+//   ['\u201F', '\u201D'],
+//   ['\u201B', '\u2019'],
+//   ['\u201E', '\u201D'],
+//   ['\u2E42', '\u201D'],
+//   ['\u201A', '\u2019'],
+//   ['\u00AB', '\u00BB'],
+//   ['\u2039', '\u203A'],
+//   ['\u2035', '\u2032'],
+//   ['\u2036', '\u2033'],
+//   ['\u2037', '\u2034'],
+//   ['\u301D', '\u301E'],
+//   ['\u0060', '\u00B4'],
+//   ['\u301F', '\u301E'],
+// ];
+//punctuation regs-  are we having fun yet?
 
-var before = /^(\s|-+|\.\.+|\/)+/;
+var before = /^([\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]|\-+|\.\.+|\/|"|"|\uFF02|'|\u201C|\u2018|\u201F|\u201B|\u201E|\u2E42|\u201A|\xAB|\u2039|\u2035|\u2036|\u2037|\u301D|`|\u301F)+/;
+var after = /([\t-\r \xA0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF]+|\-+|\.\.+|"|"|\uFF02|'|\u201D|\u2019|\u201D|\u2019|\u201D|\u201D|\u2019|\xBB|\u203A|\u2032|\u2033|\u2034|\u301E|\xB4)+$/;
 var minusNumber = /^( *)-(\$|€|¥|£)?([0-9])/;
-var after = /(\s+|-+|\.\.+)+$/;
 
 //seperate the 'meat' from the trailing/leading whitespace.
 //works in concert with ./src/text/tokenize.js
@@ -11964,6 +11955,13 @@ var methods = {
         t.whitespace.before = '';
       }
       t.whitespace.after = '';
+      //add normalized quotation symbols
+      if (t.tags.StartQuotation === true) {
+        t.whitespace.before += '"';
+      }
+      if (t.tags.EndQuotation === true) {
+        t.whitespace.after = '"' + t.whitespace.after;
+      }
     });
     return r;
   },
@@ -12642,20 +12640,37 @@ var addSubsets = function addSubsets(Text) {
       return r;
     },
     quotations: function quotations(n) {
-      var _this = this;
-
-      var quotes = [['StraightDoubleQuotes', '"', '"'], ['StraightDoubleQuotesWide', '\uFF02', '\uFF02'], ['StraightSingleQuotes', '\'', '\''], ['CommaDoubleQuotes', '\u201C', '\u201D'], ['CommaSingleQuotes', '\u2018', '\u2019'], ['CurlyDoubleQuotesReversed', '\u201F', '\u201D'], ['CurlySingleQuotesReversed', '\u201B', '\u2019'], ['LowCurlyDoubleQuotes', '\u201E', '\u201D'], ['LowCurlyDoubleQuotesReversed', '\u2E42', '\u201D'], ['LowCurlySingleQuotes', '\u201A', '\u2019'], ['AngleDoubleQuotes', '\xAB', '\xBB'], ['AngleSingleQuotes', '\u2039', '\u203A'], ['PrimeSingleQuotes', '\u2035', '\u2032'], ['PrimeDoubleQuotes', '\u2036', '\u2033'], ['PrimeTripleQuotes', '\u2037', '\u2034'], ['PrimeDoubleQuotes', '\u301D', '\u301E'], ['PrimeSingleQuotes', '`', '\xB4'], ['LowPrimeDoubleQuotesReversed', '\u301F', '\u301E']];
-      var that = null;
-      quotes.forEach(function (quote) {
-        var str = '[/.' + quote[1] + '[^' + quote[1] + ']*$/] /^[^' + quote[2] + ']*' + quote[2] + './';
-        var match = _this.match('#' + quote[0] + '+').splitAfter(str);
-        that = that === null ? match : that.concat(match);
+      var matches = this.match('#Quotation+');
+      var found = [];
+      matches.list.forEach(function (ts) {
+        var open = 0;
+        var start = null;
+        //handle nested quotes - 'startQuote->startQuote->endQuote->endQuote'
+        ts.terms.forEach(function (t, i) {
+          if (t.tags.StartQuotation === true) {
+            if (open === 0) {
+              start = i;
+            }
+            open += 1;
+          }
+          if (open > 0 && t.tags.EndQuotation === true) {
+            open -= 1;
+          }
+          if (open === 0 && start !== null) {
+            found.push(ts.slice(start, i + 1));
+            start = null;
+          }
+        });
+        //maybe we messed something up..
+        if (start !== null) {
+          found.push(ts.slice(start, ts.terms.length));
+        }
       });
-      that = that.sort('chronological');
+      matches.list = found;
       if (typeof n === 'number') {
-        that = that.get(n);
+        matches = matches.get(n);
       }
-      return that;
+      return matches;
     },
     topics: function topics(n) {
       var r = this.clauses();