+++ scratch.js
@@ -8,9 +8,9 @@ var nlp = require('./src/index');
 // console.log(doc.list[0].terms[0]);
 // console.log(doc.values().toNumber().out('text'));
 
-var doc = nlp('the so-called “fascist  dictator”');
+var doc = nlp('spencers\'');
 doc = doc.normalize();
-console.log(doc.out('text'));
+console.log(doc.out('normal'));
 doc.debug();
 // doc.quotations().debug();