+++ src/term/whitespace.js
@@ -1,8 +1,29 @@
 'use strict';
-//punctuation regs-
-const before = /^(\s|-+|\.\.+|\/)+/;
+// const quotes = [ //
+//   ['"', '"'],
+//   ['\u0022', '\u0022'],
+//   ['\uFF02', '\uFF02'],
+//   ['\u0027', '\u0027'],
+//   ['\u201C', '\u201D'],
+//   ['\u2018', '\u2019'],
+//   ['\u201F', '\u201D'],
+//   ['\u201B', '\u2019'],
+//   ['\u201E', '\u201D'],
+//   ['\u2E42', '\u201D'],
+//   ['\u201A', '\u2019'],
+//   ['\u00AB', '\u00BB'],
+//   ['\u2039', '\u203A'],
+//   ['\u2035', '\u2032'],
+//   ['\u2036', '\u2033'],
+//   ['\u2037', '\u2034'],
+//   ['\u301D', '\u301E'],
+//   ['\u0060', '\u00B4'],
+//   ['\u301F', '\u301E'],
+// ];
+//punctuation regs-  are we having fun yet?
+const before = /^(\s|-+|\.\.+|\/|"|\u0022|\uFF02|\u0027|\u201C|\u2018|\u201F|\u201B|\u201E|\u2E42|\u201A|\u00AB|\u2039|\u2035|\u2036|\u2037|\u301D|\u0060|\u301F)+/u;
+const after = /(\s+|-+|\.\.+|"|\u0022|\uFF02|\u0027|\u201D|\u2019|\u201D|\u2019|\u201D|\u201D|\u2019|\u00BB|\u203A|\u2032|\u2033|\u2034|\u301E|\u00B4)+$/u;
 const minusNumber = /^( *)-(\$|€|¥|£)?([0-9])/;
-const after = /(\s+|-+|\.\.+)+$/;
 
 //seperate the 'meat' from the trailing/leading whitespace.
 //works in concert with ./src/text/tokenize.js