+++ test/unit/subset/quotations/quotations.test.js
@@ -10,7 +10,12 @@ test('quotation test', function(t) {
     ['so \'as if\' i said', 'as if'],
     ['the \'truthiness\' i said', 'truthiness'],
     ['yeah, “fun” and stuff', 'fun'],
-    ['“Fun” and stuff', 'fun']
+    ['“Fun” and stuff', 'fun'],
+    //dangling start/end
+    ['\'twas good cookin', ''],
+    ['twas good cookin\'', ''],
+    ['twas \'good cookin\'', 'good cookin'],
+    ['\'twas \'good cookin\'', 'twas good cookin']
   ];
   arr.forEach(function(a) {
     var r = nlp(a[0]);
@@ -34,7 +39,6 @@ test('multiple quotation test', function(t) {
     [`My "String 'with manys' adjacent" "nested" 'quotes'`,
       [
         'string with manys adjacent',
-        'with manys',
         'nested',
         'quotes'
       ]
@@ -44,10 +48,8 @@ test('multiple quotation test', function(t) {
         'may\'s',
         'third day',
         'will be a really cold day',
-        'really cold',
         'in a',
         'really cold month',
-        'month'
       ]
     ],
   ];